Initial Linux-only virtual interface testing
This is an end-to-end test of rust pcap and the underlying libpcap and OS.
Currently it only works on Linux (and maybe MacOS), must be run with root
(at least CAP_NET_ADMIN) permissions.

The test creates a 'TAP' interface
(https://en.wikipedia.org/wiki/TUN/TAP) which is a virtual NIC on one
side and a file descriptor on the otherside (to fascilitate testing).

I've hidden it behind a 'tap-tests' cargo feature try to prevent
people from accidentally compile it (which requires the 'tun-tap' and
'etherparse' additional dev dependencies) or running it without meeting
the above requirements.

Current test just verifies that packets can be sent and received, e.g.,
do Capture<Active>::next_packet() and Capture<Active>::sendpacket() do
the right things.

A fair bit of work is done to avoid race conditions which can be
horrible for these types of tests.

	[robs@laptop-o7rs71ij pcap]$ cargo test -F tap-tests --no-run --test
	tap_tests | tail
	    Finished test [unoptimized + debuginfo] target(s) in 0.05s
	  Executable tests/tap_tests.rs
	(target/debug/deps/tap_tests-232532a0859d5255)
	[robs@laptop-o7rs71ij pcap]$ sudo
	./target/debug/deps/tap_tests-232532a0859d5255

	running 1 test
	test tests::conntrack_tap_basic ... ok

	test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 0 filtered
	out; finished in 0.54s