Don't automatically link to libpcap
pkg-config's default config will output `rustc-link-lib` when it finds a
package. However, per this project's readme, linkage is the
responsibility of the user, so it seems like merely probing for the
version should not change any subsequent build config.

Lines previously present in build.rs output that are now gone:

```
[pcap 2.3.0] cargo:rustc-link-search=native=/usr/lib/x86_64-linux-gnu
[pcap 2.3.0] cargo:rustc-link-lib=pcap
```

Similarly, the `#[link(name = "pcap")]` in `raw` was making dependent
binaries automatically link to libpcap.so.